move init, reset and save model to separate thread

+++ framework/core/event_manager/sources/resources/em_internal_resources.cpp
@@ -17,7 +17,7 @@ namespace Resources {
 
 /*---------------------------------------------------------------------------*/
 
-	const int TimeLimit = 16;
+	const qint64 TimeLimit = 16;
 
 	const char* const TimeLimitWarning = "Time limit has been exceeded! Time: '%1' ( limit: '%2' )";
